Manage the LdapServerResource in a HashMap
By storing the LdapServerResources in a HashMap we are providing the
groundwork for adding the ability to run and mangage multiple LDAP
server instances. As part of this change we also stop and remove any
existing servers that may have been running on a different port so that
we don't end up with running servers that we no longer have a reference
to.